requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imageio==2.37.0
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
lpips==0.1.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pip==22.0.2
platformdirs==4.3.8
portalocker==3.1.1
protobuf==6.31.1
psutil==7.0.0
ptflops==0.7.4
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.3
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools==59.6.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
tifffile==2025.5.10
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
wandb==0.20.0
yacs==0.1.8
